requirements.txt
fastapi
uvicorn
requests
llama-cpp-python
html2text
streamlit